se refactorizó character y se cambió la interfaz ICharacter por la clase abstracta
