+++ src/Library/Characters/interfaces/AbstractCharacter.cs
@@ -1,0 +1,86 @@
+using System.Collections.Generic;
+namespace RoleplayGame
+{
+    public abstract class AbstractCharacter
+    {
+        private int health;
+
+        private List<IItem> items;
+
+        public AbstractCharacter(string name)
+        {
+            this.Name = name;
+            this.items = new List<IItem>();
+            this.Health = 100;
+        }
+
+        protected string Name { get; set; }
+        
+        protected int AttackValue
+        {
+            get
+            {
+                int value = 0;
+                foreach (IItem item in this.items)
+                {
+                    if (item is IAttackItem)
+                    {
+                        value += (item as IAttackItem).AttackValue;
+                    }
+                }
+                return value;
+            }
+        }
+
+        protected int DefenseValue
+        {
+            get
+            {
+                int value = 0;
+                foreach (IItem item in this.items)
+                {
+                    if (item is IDefenseItem)
+                    {
+                        value += (item as IDefenseItem).DefenseValue;
+                    }
+                }
+                return value;
+            }
+        }
+
+        public int Health
+        {
+            get
+            {
+                return this.health;
+            }
+            private set
+            {
+                this.health = value < 0 ? 0 : value;
+            }
+        }
+
+        public void ReceiveAttack(int power)
+        {
+            if (this.DefenseValue < power)
+            {
+                this.Health -= power - this.DefenseValue;
+            }
+        }
+
+        public void Cure()
+        {
+            this.Health = 100;
+        }
+
+        public void AddItem(IItem item)
+        {
+            this.items.Add(item);
+        }
+
+        public void RemoveItem(IItem item)
+        {
+            this.items.Remove(item);
+        }
+    }
+}
