+++ src/Library/Characters/interfaces/IMagicCharacter.cs
@@ -1,6 +1,6 @@
 namespace RoleplayGame
 {
-    public interface IMagicCharacter: ICharacter
+    public interface IMagicCharacter
     {
         void AddItem(IMagicalItem item);
