+++ src/Library/Items/interfaces/IDefenseItem.cs
