+++ src/Library/Items/interfaces/IItem.cs
