+++ src/Library/Items/interfaces/IMagicalAttackItem.cs
