+++ src/Library/Items/interfaces/IMagicalDefenseItem.cs
