+++ src/Library/Items/interfaces/IMagicalItem.cs
