+++ src/Library/Items/interfaces/ISpell.cs
