--- src/test/Test.Library/UnitTest1.cs
@@ -1,18 +1,0 @@
-using NUnit.Framework;
-
-namespace Test.Library
-{
-    public class Tests
-    {
-        [SetUp]
-        public void Setup()
-        {
-        }
-
-        [Test]
-        public void Test1()
-        {
-            Assert.Pass();
-        }
-    }
-}
